Update and rename connect.go to main.go
